A few more

+++ lib/lib-javadoc/src/test/java/org/sagebionetworks/javadoc/BasicFileUtilsTest.java
@@ -34,7 +34,9 @@ public class BasicFileUtilsTest {
 			String results = BasicFileUtils.pathToRoot(temp.getParentFile(), child);
 			assertEquals("../../", results);
 		}finally{
-			temp.delete();
+			if (temp != null) {
+				temp.delete();
+			}
 		}
 	}
